requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
geojson==3.0.1
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.0
psycopg2==2.9.6
python-dotenv==1.0.0
shapely==2.0.1
Werkzeug==2.3.6
zipp==3.15.0
